Adjusting DDG-DA
